fix(client): Phase 12 CR cleanup — address 13 promotion-PR findings
Batch-fixes the 14 CR findings raised on promotion PR #71 except
the new feature add (ExposureControlsPanel filter-slot UI) which is
tracked in design/PORT_TODO.md for a Phase 12c follow-up.

3 Critical — clamp() type:
- wizard_state.dart jumpTo .toInt() on clamp result
- histogram_strip.dart alpha .toDouble()
- frame_quality_chart.dart yMax .toDouble()

4 Major:
- wizard_shell.dart: Save & Exit always dismisses the route
- saved_server_state.dart: in-memory state updates on persist
  failure so first-run isn't blocked by transient keyring errors
- sequence_state.dart: addChild + addSiblingAfter refuse
  SequenceNodeKind.root (single-root invariant)
- calendar_heatmap.dart: real DateTime.now() anchor, not hardcoded

6 Minor:
- framing_tab Set as Target button disabled until 12c.2 wires it
- imaging_tab status pill tap handler now null
- equipment_chip semantics: button = onTap OR onLongPress
- sequencer_toolbar buttons in SingleChildScrollView for narrow
  windows
- guiding_rms_chart axis title "session #" → "Session date"
- PORT_PROGRESS.md text reconciled

+++ client/openastroara_client/lib/state/wizard_state.dart
@@ -34,7 +34,9 @@ class WizardController extends Notifier<WizardState> {
   }
 
   void jumpTo(int step) {
-    final clamped = step.clamp(1, ProfileWizard.totalSteps);
+    // `clamp` returns `num`; copyWith expects `int`. Explicit toInt() so
+    // static analysis is happy.
+    final clamped = step.clamp(1, ProfileWizard.totalSteps).toInt();
     state = state.copyWith(step: clamped);
   }
