+++ client/openastroara_client/lib/widgets/imaging/histogram_strip.dart
@@ -21,7 +21,8 @@ class HistogramStrip extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.end,
         children: List<Widget>.generate(64, (i) {
           // Placeholder gradient — empty bins from left to right.
-          final alpha = (0.05 + 0.25 * (i / 63)).clamp(0.0, 1.0);
+          // `clamp` returns `num`; withValues alpha expects `double`.
+          final alpha = (0.05 + 0.25 * (i / 63)).clamp(0.0, 1.0).toDouble();
           return Expanded(
             child: Container(
               margin: const EdgeInsets.symmetric(horizontal: 0.5),
